+++ lib/src/structure_creators/file/impl_file_creator.dart
@@ -14,7 +14,9 @@ class ImplFileCreator implements IFileCreator {
     await _createFile(
       directoryCreator.constantDir.path,
       'app_url',
-      content: """class AppUrl{ static const BASE_URL = "";}""",
+      content: """class AppUrl {
+  static const BASE_URL = "";
+}""",
     );
     await _createFile(
       directoryCreator.constantDir.path,
@@ -30,7 +32,6 @@ import 'package:dio/dio.dart' as d;
 import 'package:riverpod_test/constant/app_url.dart';
 import 'package:riverpod_test/constant/constant_key.dart';
 import 'package:riverpod_test/data_provider/pref_helper.dart';
-import 'package:riverpod_test/mvc/get_module/views/get_screen.dart';
 import 'package:riverpod_test/utils/enum.dart';
 import 'package:riverpod_test/utils/navigation_service.dart';
 import 'package:riverpod_test/utils/view_util.dart';
@@ -128,8 +129,8 @@ class ApiClient {
 
             switch (code) {
               case 401:
-                PrefHelper.setString(TOKEN, "").then((value) => GetScreen()
-                    .pushAndRemoveUntil(Navigation.key.currentContext));
+                // PrefHelper.setString(TOKEN, "").then((value) => GetScreen()
+                //     .pushAndRemoveUntil(Navigation.key.currentContext));
 
                 break;
               default:
@@ -190,6 +191,203 @@ class ApiClient {
   }
 }
 """);
+    await _createFile(
+      directoryCreator.dataProviderDir.path,
+      'graph_client',
+      content: """import 'dart:convert';
+import 'dart:io';
+import 'package:dio/dio.dart' as d;
+import 'package:flutter/material.dart';
+import 'package:riverpod_test/constant/app_url.dart';
+import 'package:riverpod_test/constant/constant_key.dart';
+import 'package:riverpod_test/data_provider/pref_helper.dart';
+import 'package:riverpod_test/global/model/graph_ql_error_response.dart';
+import 'package:riverpod_test/utils/navigation_service.dart';
+import 'package:riverpod_test/utils/view_util.dart';
+
+class ApiClient {
+  late d.Dio _dio;
+
+  Map<String, dynamic> _header = {};
+
+  _initDio({String baseUrl = AppUrl.BASE_URL}) {
+    _header = {
+      HttpHeaders.contentTypeHeader: "application/json",
+      HttpHeaders.authorizationHeader: "Bearer \${PrefHelper.getString(TOKEN)}"
+    };
+
+    _dio = d.Dio(
+      d.BaseOptions(
+        baseUrl: baseUrl,
+        headers: _header,
+        connectTimeout: 1000 * 30,
+        sendTimeout: 1000 * 10,
+      ),
+    );
+    _initInterceptors();
+  }
+
+  void _initInterceptors() {
+    _dio.interceptors.add(d.InterceptorsWrapper(onRequest: (options, handler) {
+      print(
+          'REQUEST[\${options.method}] => PATH: \${AppUrl.BASE_URL}\${options.path} '
+          '=> Request Values: param: \${options.queryParameters}, DATA: \${options.data}, => HEADERS: \${options.headers}');
+      return handler.next(options);
+    }, onResponse: (response, handler) {
+      print(
+          'RESPONSE[\${response.statusCode}] => DATA: \${response.data} URL: \${response.requestOptions.baseUrl}\${response.requestOptions.path}');
+      return handler.next(response);
+    }, onError: (err, handler) {
+      print(
+          'ERROR[\${err.response?.statusCode}] => DATA: \${err.response?.data} Message: \${err.message} URL: \${err.response?.requestOptions.baseUrl}\${err.response?.requestOptions.path}');
+      return handler.next(err);
+    }));
+  }
+
+//This requestFormData is actually for REST API
+//Usign this we sent multipart file thats why its argument
+//is different in request method
+  Future requestFormData(String url, Map<String, File>? files,
+      {bool isLoaderShowing = false}) async {
+    try {
+      if (isLoaderShowing) CircularProgressIndicator();
+
+      d.Response response;
+      _header[d.Headers.contentTypeHeader] = 'multipart/form-data';
+      _initDio(baseUrl: AppUrl.FILE_UPLOAD_BASE_URL);
+
+      Map<String, d.MultipartFile> fileMap = {};
+      if (files != null) {
+        for (MapEntry fileEntry in files.entries) {
+          File file = fileEntry.value;
+          fileMap[fileEntry.key] = await d.MultipartFile.fromFile(file.path);
+        }
+      }
+
+      Map<String, dynamic> params = Map<String, dynamic>();
+      params.addAll(fileMap);
+      final data = d.FormData.fromMap(params);
+
+      // Handle response code from api
+      response = await _dio.post(url, data: data);
+
+      if (isLoaderShowing) Navigation.pop(Navigation.key.currentContext);
+
+      if (response.statusCode == 200) {
+        return response;
+      } else {
+        _showExceptionSnackBar("Something went wrong");
+        throw Exception();
+      }
+
+      // Handle Error type if dio catches anything
+    } on d.DioError catch (e) {
+      _dioErrorHandler(isLoaderShowing, e);
+    } catch (e) {
+      throw Exception();
+    }
+  }
+
+  //GraphQL client using dio for Mutation and query
+  Future request(
+      {String body = "",
+      String url = '',
+      Map<String, dynamic> variables = const {},
+      bool isLoaderShowing = false}) async {
+    if (isLoaderShowing) CircularProgressIndicator();
+
+    d.Response response;
+    _initDio(baseUrl: AppUrl.BASE_URL);
+
+    try {
+      String paramJson = ''' 
+      {
+      "query" : "\$body",
+      "variables" : \${jsonEncode(variables)}
+      }
+      '''
+          .replaceAll("\n", "");
+
+      response = await _dio.post(url, data: paramJson);
+
+      if (isLoaderShowing) Navigation.pop(Navigation.key.currentContext);
+
+      if (response.statusCode == 200) {
+        if (response.data['errors'] != null) {
+          handleGraphQlError(response);
+        } else {
+          return response;
+        }
+      } else {
+        _showExceptionSnackBar("Something went wrong");
+        throw Exception();
+      }
+
+      // Handle Error type if dio catches anything
+    } on d.DioError catch (e) {
+      _dioErrorHandler(isLoaderShowing, e);
+    } catch (e) {
+      throw Exception();
+    }
+  }
+
+  void handleGraphQlError(d.Response response) {
+    try {
+      final result = GraphQLErrorResponse.fromJson(response.data);
+
+      if (result != null) {
+        if (result.errors.isNotEmpty) {
+          if (result.errors[0].message == "Invalid User" ||
+              result.errors[0].message == "User does not exist") {
+            PrefHelper.logout();
+           // GetScreen().pushAndRemoveUntil(Navigation.key.currentContext);
+          } else
+            _showExceptionSnackBar(result.errors[0].message);
+        }
+      }
+    } catch (e) {}
+  }
+
+// Handle Error type if dio catches anything
+  void _dioErrorHandler(bool isLoaderShowing, d.DioError dioError) {
+    if (isLoaderShowing) Navigation.pop(Navigation.key.currentContext);
+
+    switch (dioError.type) {
+      case d.DioErrorType.response:
+        if (dioError.response != null) {
+          if (dioError.response!.statusCode != null) {
+            if (dioError.response!.statusCode! == 500) {
+              _showExceptionSnackBar("Server Error");
+            }
+          }
+        }
+
+        break;
+      case d.DioErrorType.connectTimeout:
+        _showExceptionSnackBar("Connection failed!. Please refresh");
+        // return request(body: body, variables: variables);
+        break;
+      case d.DioErrorType.sendTimeout:
+        // return request(body: body, variables: variables);
+        break;
+      case d.DioErrorType.receiveTimeout:
+        break;
+      case d.DioErrorType.other:
+        if (dioError.error is SocketException) {
+          _showExceptionSnackBar("Check your internet connection");
+        }
+        break;
+      default:
+        _showExceptionSnackBar("Something went wrong!");
+    }
+  }
+
+  static _showExceptionSnackBar(String? msg) async {
+    ViewUtil.SSLSnackbar(msg ?? "");
+  }
+}
+ """,
+    );
 
     await _createFile(directoryCreator.dataProviderDir.path, 'pref_helper.dart',
         content: """import 'package:riverpod_test/constant/constant_key.dart';
@@ -267,6 +465,48 @@ class PrefHelper {
 }
 """);
 
+//global model folder
+    await _createFile(
+      directoryCreator.globalDir.path + '/model',
+      'graph_ql_error_response',
+      content: """class GraphQLErrorResponse {
+  GraphQLErrorResponse({
+    required this.errors,
+  });
+
+  late final List<Errors> errors;
+
+  GraphQLErrorResponse.fromJson(Map<String, dynamic> json) {
+    errors = List.from(json['errors']).map((e) => Errors.fromJson(e)).toList();
+  }
+
+  Map<String, dynamic> toJson() {
+    final _data = <String, dynamic>{};
+    _data['errors'] = errors.map((e) => e.toJson()).toList();
+    return _data;
+  }
+}
+
+class Errors {
+  Errors({
+    required this.message,
+  });
+
+  late final String message;
+
+  Errors.fromJson(Map<String, dynamic> json) {
+    message = json['message'];
+  }
+
+  Map<String, dynamic> toJson() {
+    final _data = <String, dynamic>{};
+    _data['message'] = message;
+    return _data;
+  }
+}
+""",
+    );
+
     //localization file
     await _createFile(
       directoryCreator.l10nDir.path,
@@ -318,26 +558,133 @@ class PrefHelper {
 
 //Utils file
     await _createFile(directoryCreator.utilsDir.path + '/styles', 'k_assets',
-        content:
-            """class KAssets { static const _rootPath = 'assets'; static const _svgDir = '\$_rootPath/svg'; static const _imageDir = '\$_rootPath/images'; }""");
+        content: """class KAssets {
+  static const _rootPath = 'assets';
+  static const _svgDir = '\$_rootPath/svg';
+  static const _imageDir = '\$_rootPath/images';
+}
+""");
     await _createFile(directoryCreator.utilsDir.path + '/styles', 'k_colors',
-        content:
-            """import 'package:flutter/material.dart'; class KColors { static const primary = Color(0xFF299E8D); static const accent = Color(0x1F299E8D); static const darkAccent = Color(0xFF2C4251); static const white = Colors.white; static const black = Colors.black; static final charcoal = Color(0xFF264654); static final lightCharcoal = charcoal.withOpacity(.12); static const spaceCadet = Color(0xFF2C3549); static final lightRed = Colors.red[100]; static final red = Colors.red; static final transparent = Colors.transparent; }""");
+        content: """import 'package:flutter/material.dart';
+
+class KColors {
+  static const primary = Color(0xFF299E8D);
+  static const accent = Color(0x1F299E8D);
+  static const darkAccent = Color(0xFF2C4251);
+  static const white = Colors.white;
+  static const black = Colors.black;
+  static final charcoal = Color(0xFF264654);
+  static final lightCharcoal = charcoal.withOpacity(.12);
+  static const spaceCadet = Color(0xFF2C3549);
+  static final lightRed = Colors.red[100];
+  static final red = Colors.red;
+  static final transparent = Colors.transparent;
+}
+""");
     await _createFile(directoryCreator.utilsDir.path + '/styles', 'k_size',
-        content:
-            """import 'package:flutter/material.dart'; import 'package:riverpod_test/utils/navigation_service.dart'; \n//zeplin size \n // width 414 \n// height 896 \nextension KSizes on num { static Size get screenSize => MediaQuery.of(Navigation.key.currentContext!).size; \n//height \ndouble get h => (this / 896) * (screenSize.height > 896 ? 896 : screenSize.height); \n//Width\n double get w => (this / 414) * (screenSize.width > 414 ? 414 : screenSize.width); \n//fontSize\n double get sp {\n // For small devices. \nif (screenSize.height < 600) { return 0.7 * this; }\n // \nFor normal device return 1.0 * this; } }""");
+        content: """import 'package:flutter/material.dart';
+import 'package:riverpod_test/utils/navigation_service.dart';
+
+//zeplin size
+// width 414
+// height 896
+extension KSizes on num {
+  static Size get screenSize =>
+      MediaQuery.of(Navigation.key.currentContext!).size;
+
+  //height
+  double get h =>
+      (this / 896) * (screenSize.height > 896 ? 896 : screenSize.height);
+
+  //Width
+  double get w =>
+      (this / 414) * (screenSize.width > 414 ? 414 : screenSize.width);
+
+  //fontSize    
+  double get sp {
+    // For small devices.
+    if (screenSize.height < 600) {
+      return 0.7 * this;
+    }
+    // For normal device
+    return 1.0 * this;
+  }
+}
+""");
     await _createFile(
         directoryCreator.utilsDir.path + '/styles', 'k_text_style',
-        content:
-            """import 'package:flutter/material.dart'; import 'package:google_fonts/google_fonts.dart'; import 'package:riverpod_test/utils/styles/styles.dart'; class KTextStyle { static TextStyle headLine3 = GoogleFonts.quicksand( fontSize: 42.sp, fontWeight: FontWeight.w500, ); static TextStyle headLine4 = GoogleFonts.quicksand( fontSize: 32.sp, fontWeight: FontWeight.w500, ); static TextStyle buttonText({fontWeight = FontWeight.normal}) => GoogleFonts.quicksand( fontSize:27.sp, fontWeight: fontWeight, ); \n/// Normal Texts static TextStyle bodyText1() => GoogleFonts.quicksand( fontSize: 27.sp, fontWeight: FontWeight.normal, ); static TextStyle bodyText2() => GoogleFonts.quicksand( fontSize: 24.sp, fontWeight: FontWeight.w500, ); static TextStyle bodyText3() => GoogleFonts.quicksand( fontSize: 22.sp, fontWeight: FontWeight.normal, );\N /// Subtitles\n static TextStyle subtitle1 =const TextStyle( fontSize: 16, fontWeight: FontWeight.w400, letterSpacing: 0.15, ); static TextStyle subtitle2 =const TextStyle( fontSize: 14, fontWeight: FontWeight.w500, letterSpacing: 0.1, ); }""");
+        content: """import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:riverpod_test/utils/styles/styles.dart';
+
+class KTextStyle {
+  static TextStyle headLine3 = GoogleFonts.quicksand(
+    fontSize: 42.sp,
+    fontWeight: FontWeight.w500,
+  );
+
+  static TextStyle headLine4 = GoogleFonts.quicksand(
+    fontSize: 32.sp,
+    fontWeight: FontWeight.w500,
+  );
+
+  static TextStyle buttonText({fontWeight = FontWeight.normal}) => GoogleFonts.quicksand(
+        fontSize:27.sp,
+        fontWeight: fontWeight,
+      );
+
+  /// Normal Texts
+  static TextStyle bodyText1() => GoogleFonts.quicksand(
+        fontSize: 27.sp,
+        fontWeight: FontWeight.normal,
+      );
+
+  static TextStyle bodyText2() => GoogleFonts.quicksand(
+        fontSize: 24.sp,
+        fontWeight: FontWeight.w500,
+      );
+
+  static TextStyle bodyText3() => GoogleFonts.quicksand(
+        fontSize: 22.sp,
+        fontWeight: FontWeight.normal,
+      );
+
+  /// Subtitles
+  static TextStyle subtitle1 =const TextStyle(
+    fontSize: 16,
+    fontWeight: FontWeight.w400,
+    letterSpacing: 0.15,
+  );
+
+  static TextStyle subtitle2 =const TextStyle(
+    fontSize: 14,
+    fontWeight: FontWeight.w500,
+    letterSpacing: 0.1,
+  );
+}
+""");
 
     await _createFile(directoryCreator.utilsDir.path + '/styles', 'k_assets',
-        content:
-            """export 'k_colors.dart'; export 'k_size.dart'; export 'k_text_style.dart'; export 'k_assets.dart'""");
+        content: """export 'k_colors.dart';
+export 'k_size.dart';
+export 'k_text_style.dart';
+export 'k_assets.dart';
+""");
 
     await _createFile(directoryCreator.utilsDir.path, 'enum',
-        content:
-            """enum LanguageOption { Bangla, English, } enum CART_STATUS { INCREMENT, REMOVE, DECREMENT, } enum Method { POST, GET, PUT, DELETE, PATCH }""");
+        content: """enum LanguageOption {
+  Bangla,
+  English,
+}
+
+enum CART_STATUS {
+  INCREMENT,
+  REMOVE,
+  DECREMENT,
+}
+
+enum Method { POST, GET, PUT, DELETE, PATCH }""");
+
     await _createFile(directoryCreator.utilsDir.path, 'extention',
         content: """import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:riverpod_test/data_provider/pref_helper.dart';
@@ -526,7 +873,26 @@ extension Navigation on Widget {
 """);
     await _createFile(directoryCreator.utilsDir.path, 'view_util',
         content:
-            """import 'package:flutter/material.dart'; import 'package:riverpod_test/utils/navigation_service.dart'; class ViewUtil { static SSLSnackbar(String msg) {\n //Using ScaffoldMessenger we can easily access \n//this snackbar from anywhere\n return ScaffoldMessenger.of(Navigation.key.currentContext!).showSnackBar( SnackBar( content: Text(msg), action: SnackBarAction( label: '', textColor: Colors.transparent, onPressed: () {}, ), ), ); } }""");
+            """import 'package:flutter/material.dart';
+import 'package:riverpod_test/utils/navigation_service.dart';
+
+class ViewUtil {
+  static SSLSnackbar(String msg) {
+    //Using ScaffoldMessenger we can easily access
+//this snackbar from anywhere
+    return ScaffoldMessenger.of(Navigation.key.currentContext!).showSnackBar(
+      SnackBar(
+        content: Text(msg),
+        action: SnackBarAction(
+          label: '',
+          textColor: Colors.transparent,
+          onPressed: () {},
+        ),
+      ),
+    );
+  }
+}
+""");
     await _createFile(
       'lib',
       'main',
@@ -535,8 +901,6 @@ import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:riverpod_test/data_provider/pref_helper.dart';
-import 'package:riverpod_test/mvc/get_module/views/get_screen.dart';
-import 'package:riverpod_test/mvc/screen_dynamic/screen_dynamic.dart';
 import 'package:riverpod_test/utils/navigation_service.dart';
 import 'package:riverpod_test/utils/styles/styles.dart';
 //localization
@@ -563,7 +927,7 @@ class MyApp extends StatelessWidget {
     SystemChannels.textInput.invokeMethod('TextInput.hide');
 
     return MaterialApp(
-      title: 'SSL template',
+      title: '',
       navigatorKey: Navigation.key,
       debugShowCheckedModeBanner: false,
 //localization
@@ -592,7 +956,7 @@ class MyApp extends StatelessWidget {
           ),
         ),
       ),
-      home: GetScreen(),
+      home: ,
     );
   }
 }